Separate classes to files
