+++ tictactoe/player.rb
@@ -1,0 +1,8 @@
+class Player
+  attr_reader :name, :symbol
+
+  def initialize(name, symbol)
+    @name = name
+    @symbol = symbol
+  end
+end
